+++ CHANGELOG.md
@@ -1529,3 +1529,8 @@ All notable changes to this project will be documented in this file.
 ## [3.6.14] - 04.05.2025
 
 - use compression for large message related to in-game debugging
+
+## [3.6.15] - 04.05.2025
+
+- wait for context disposal when stopping in order to get process properly killed in game
+- listen for exit, SIGINT, SIGUSR1, SIGUSR2, SIGTERM and uncaughtException event to dispose context